Previously, only SYSTEM user is prevented from preauthentication, by emulating root if the uid matches the RID of the regular SYSTEM user. Every other owner of the imapd process is considered non-root and therefore, preauthenticated.
From Windows 2003 Server (I think), Cygwin has had to create an
alternative SYSTEM user, with CreateToken privilege. Usually, it is
called cyg_server, and the csih helper scripts (ssh-host-config,
iu-config, etc.) use this as the default user to start the Windows
service under.

This meant that on those systems, imapd would consider this non-SYSTEM
user as preauthenticated. This would mean that any imapd connection
would not be authenticated and automatically pointed at /var/empty as
the home directory containing imap folders.

This patch resolves this in an not so direct way. The intention of the
original imap code is to disallow root/privileged users from being
preauthenticated. So this patch also (to SYSTEM user) disallows
preauth (by emulating root) if the process owner (cyg_server) is in
the Administrators group. Again, it checks against the RID of the
regular Administrators group, although this can also be circumvented
with /etc/group mods. Small price to pay for simplicity.